Web-HW13/pyproject.toml
[tool.poetry]
name = "web-hw11"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.1"
pydantic-settings = "^2.2.1"
sphinx = "^7.2.6"
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
